+++ nova/modules/core/views/_base/main/js/sim_stats_js.php
@@ -3,5 +3,7 @@
 <script type="text/javascript">
 	$(document).ready(function(){
 		$('table.zebra tbody > tr:nth-child(odd)').addClass('alt');
+
+		$('#tabs').tabs();
 	});
 </script>